Move keyrings to interfaces.ts
