+++ components/brave_wallet_ui/common/hardware/interfaces.ts
